Move helper to util
